Fix warning: implicit conversion changes signedness
visp/modules/sensor/test/rgb-depth/testRealSense2_SR300.cpp: In function ‘int main(int, char**)’:
visp/modules/sensor/test/rgb-depth/testRealSense2_SR300.cpp:348:27: warning: conversion to ‘uint32_t’ {aka ‘unsigned int’} from ‘size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  348 |       pointcloud->width = static_cast<size_t>(depth_profile.width());
      |                           ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/modules/sensor/test/rgb-depth/testRealSense2_SR300.cpp:349:28: warning: conversion to ‘uint32_t’ {aka ‘unsigned int’} from ‘size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  349 |       pointcloud->height = static_cast<size_t>(depth_profile.height());
      |                            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ modules/sensor/test/rgb-depth/testRealSense2_SR300.cpp
@@ -345,8 +345,8 @@ int main(int argc, char *argv[])
     {
       std::lock_guard<std::mutex> lock(mutex);
 
-      pointcloud->width = static_cast<size_t>(depth_profile.width());
-      pointcloud->height = static_cast<size_t>(depth_profile.height());
+      pointcloud->width = static_cast<uint32_t>(depth_profile.width());
+      pointcloud->height = static_cast<uint32_t>(depth_profile.height());
       pointcloud->points.resize(pointcloud_colvector.size());
       for (size_t i = 0; i < pointcloud_colvector.size(); i++) {
         pointcloud->points[static_cast<size_t>(i)].x = pointcloud_colvector[i][0];
